pyproject.toml
[project]
name = "abuddy"
version = "0.1.0"
description = "AWS Certified Generative AI Developer Professional study app"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "boto3>=1.35.0",
    "orjson>=3.10.0",
    "loguru>=0.7.2",
    "networkx>=3.4.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "PyJWT[crypto]>=2.9.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/abuddy"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
